move lepton_header into its own module
