+++ src/structs/jpeg_read.rs
@@ -41,7 +41,7 @@ use crate::helpers::here;
 use super::bit_reader::BitReader;
 use super::block_based_image::{AlignedBlock, BlockBasedImage};
 use super::jpeg_position_state::JpegPositionState;
-use super::lepton_format::LeptonHeader;
+use super::lepton_header::LeptonHeader;
 use super::thread_handoff::ThreadHandoff;
 use crate::lepton_error::ExitCode;
